Fix: SSL shutdown phase could block.
During SSL_shutdown() call an nrpe child process can be blocked because SSL_shutdown() needs to do a read on the socket
but the connection could have been broken in a way or another (ESTABLISHED on this side, disappeared on other one).

As nrpe uses blocking socket mode -> process blocked.
Prevent this with an alarm signal handler.