Return UNKNOWN when issues occur
Discovered a few instances where poorly written plugins will return an
OK status to NRPE with no output. This results in NRPE returning the
message "NRPE: Unable to read output" and an OK status. When a plugin
doesn't return any output at all this should be considered an issue.

Also, when a command run by NRPE times out it should also be handled as
an UNKNOWN situation. In most cases, check_nrpe will timeout first and
however it is configured will decide on the status returned. However,
just to be in the safe side, should NRPE return data before check_nrpe
hits a it's timeout, return an UNKNOWN state when a timeout is
encountered.