Additional SSL/TLS Changes
Added the ability to use hex values for the ssl_logging directive
in nrpe.cfg to make it easier to combine multiple values. Plus a
couple of fairly minor fixes and corrections to nrpe.c

In check_nrpe.c, added a logging argument and associated calls to
syslog, and changed the -n option to optionally take a number as
an argument to match the way ADH is handled in nrpe.

Updated the README.SSL.md file for the updated command-line
arguments in check_nrpe.