check_nrpe - config file might be handy for newer SSL stuff
fix for issue: https://github.com/NagiosEnterprises/nrpe/issues/34

A config file can now be specified for check_nrpe using a '-f'
or --config-file= command-line option. The config file may NOT
include any of the following options:
    --config_file (-f), --command (-c), --args (-a)
An error will be logged to syslog if they exist.

Other options given on the command-line will override any existing
options in the config file, and a warning will be logged to syslog.

Also updated the "Last Modified" date in nrpe.cfg.in