nrpe_user directive doesn't work as expected?
Fix for issue https://github.com/NagiosEnterprises/nrpe/issues/40

I had changed the call to setuid() to seteuid() so the PID file
could be deleted. Unfortunately, when a popen() is done, some shells
(like bash which is a common default shell) reverts to the original
uid. So I added a call go setuid() in my_system() after the fork().

Also updated Changelog and added a missing #ifdef HAVE_SSL