Update nrpe.cfg.in
The wload and cload for check_load should be specified in decimal notation. The current defaults in percent notation would never work correctly in the latest versions of this plugin, and they're - unfortunately - probably the basis for a lot of people's production configuration. See https://github.com/nagios-plugins/nagios-plugins/blob/master/plugins/check_load.c#L200 for why percent notation doesn't work.

I would also add "-r" to divide by the number of cores, as that provides the most user-friendly way to specify wload and cload without knowing the number of CPUs on a system and without adding explanatory notes to the nrpe.cfg file.