Fixes in parse_allowed_hosts() and called functions
Fix for issues:
https://github.com/NagiosEnterprises/nrpe/issues/54
https://github.com/NagiosEnterprises/nrpe/issues/55
https://github.com/NagiosEnterprises/nrpe/issues/56

Thanks to Jobst Schmalenbach for finding one of the problems in the acl.c
file. First of all, `config.h` was not being included so `strtok_r` was
never used in place of `strtok`. Secondly, on systems that don't have
`strtok_r`, the call to `add_ipv6_to_acl` includes calls to `strtok`
that overwrote the internal context. So if a call to `add_ipv4_to_acl`
failed because it was a domain name or an IPv6 address, the `strtok` loop
in `parse_allowed_hosts` returned NULL and any further hosts would be
completely skipped.

I changed the `add_ipv6_to_acl` function to not use `strtok` or
`strtok_r`, and instead used `strchr` to look for a '/' for the
mask and manually break the string apart.