Getting "Unrecognized option" if parameters after `-a` contain dashes
If the command ended in something like `-a --foo --bar baz`,
getopt_long would complain about `--bar` but check_nrpe would pass
along all the arguments anyway. The check for `-a` swallowed the
`--foo`, but the `--bar` confused it because the loop check after
the call to getopt_long for `if (argindex > 0)` instead of before.