"Remote %s accepted a Version %s Packet", please add to debug
Partial fix for issue #72

I would really rather not add another command-line flag at this
point, so instead I changed a little what gets logged.

If check_nrpe is v3.x, it will NOT log any messages if the remote
accepts a version 3 packet. If the `-2` switch (only use version 2
packet) was set, it will NOT log any messages. It will only log
a message if it first tried to send a version 3 packet, failed,
and sent a version 2 packet which succeeded. It will be logged
with logleve `LOG_DEBUG`.

It's not quite what the origianal poster wanted, but it should
reduce the number of log messages.