nrpe 3.0.1 segfaults when key and/or cert are broken symlinks
Fix for issue #76

This seems to be somewhat platform specific. I was unable to
reproduce the problem on an openSUSE VM, but the segfault did
happen on a CentOS 5 VM.

The `ERR_error_string()` call seemed to be passing back an invalid
char pointer whether or not the second parameter was NULL. So I
put in a 120 character buffer and printed the error message from
there, instead of the returned pointer.