Allow NRPE logging directly rather than requiring use of syslog
Fix for issue #65

In the 'nrpe.cfg' file, there is now an option 'log-file=<path>'. When this
option is set, everything that would normally get logged to syslog will now
be written to the specified file.

Similarly, 'check_nrpe' has the long command line option '--log-file' or
the short option '-g'. This option can also be specified in the config file
given in the '--config-file' ('-f') parameter.