OpenSSL 1.1.0e: Error: Could not complete SSL handshake
(sort of) fix for issue #119

Anonymous Diffie Hellman is not available in OpenSSL 1.1.x so
if there's no certificates, it ends up with a "no shared cipher"
error. It does the same thing if you specify `-d0` on the
`check_nrpe` command line, which disables ADH.

Logging for both nrpe and check_nrpe has been modified to include
more information. For this issue in particular, the nrpe daemon
checks for "no shared cipher" and puts out an additional log line
stating: "Error: This could be because you have not specified
certificate or ca-certificate files".

So it's not really fixed (and it can't be) but the problem is
better documented.