startup/openrc-init.in: supply a default value for NRPE_CFG.
The OpenRC init script requires the NRPE_CFG variable to be set. The
startup/openrc-conf.in file sets it, but that file is not installed as
part of the usual "make install-init" process. To ensure that the init
script works out of the box, this commit sets a default value for
NRPE_CFG directly in the init script.

Closes: https://github.com/NagiosEnterprises/nrpe/issues/165